+++ src/Cooked/MockChain.hs
@@ -6,6 +6,7 @@ import Cooked.MockChain.Balancing as X
 import Cooked.MockChain.BlockChain as X hiding (MockChainLogEntry, logEvent)
 import Cooked.MockChain.Direct as X hiding (MockChainReturn)
 import Cooked.MockChain.MinAda as X
+import Cooked.MockChain.MockChainSt as X (MockChainSt (..), mockChainSt0From)
 import Cooked.MockChain.Staged as X hiding (StagedMockChain)
 import Cooked.MockChain.Testing as X
 import Cooked.MockChain.UtxoSearch as X