+++ src/Cooked/MockChain/Balancing.hs
@@ -78,7 +78,11 @@ balanceTxSkel skelUnbal@TxSkel {..} = do
     -- We retrieve the various kinds of scripts
     spendingScripts <- txSkelInputValidators skelUnbal
     -- The transaction will only require collaterals when involving scripts
-    let noScriptInvolved = Map.null txSkelMints && null (mapMaybe txSkelProposalWitness txSkelProposals) && Map.null spendingScripts
+    let noScriptInvolved =
+          Map.null txSkelMints
+            && null (mapMaybe txSkelProposalWitness txSkelProposals)
+            && Map.null spendingScripts
+            && null (txSkelWithdrawalsScripts skelUnbal)
     case (noScriptInvolved, txOptCollateralUtxos txSkelOpts) of
       (True, CollateralUtxosFromSet utxos _) -> logEvent (MCLogUnusedCollaterals $ Right utxos) >> return Nothing
       (True, CollateralUtxosFromWallet cWallet) -> logEvent (MCLogUnusedCollaterals $ Left cWallet) >> return Nothing
@@ -298,7 +302,7 @@ estimateTxSkelFee :: (MonadBlockChainBalancing m) => TxSkel -> Fee -> Maybe (Col
 estimateTxSkelFee skel fee mCollaterals = do
   -- We retrieve the necessary data to generate the transaction body
   params <- getParams
-  managedData <- txSkelInputData skel
+  managedData <- txSkelHashedData skel
   let collateralIns = case mCollaterals of
         Nothing -> []
         Just (s, _) -> Set.toList s
@@ -310,7 +314,7 @@ estimateTxSkelFee skel fee mCollaterals = do
     Right txBodyContent -> return txBodyContent
   -- We create the actual body and send if for validation
   txBody <- case Cardano.createAndValidateTransactionBody Cardano.ShelleyBasedEraConway txBodyContent of
-    Left err -> throwError $ MCEGenerationError (TxBodyError "Error creating body when estimating fees" err)
+    Left err -> throwError $ MCEGenerationError $ TxBodyError "Error creating body when estimating fees" err
     Right txBody -> return txBody
   -- We retrieve the estimate number of required witness in the transaction
   let nkeys = Cardano.estimateTransactionKeyWitnessCount txBodyContent
@@ -331,24 +335,40 @@ estimateTxSkelFee skel fee mCollaterals = do
 
 -- | This creates a balanced skeleton from a given skeleton and fee. In other
 -- words, this ensures that the following equation holds: input value + minted
--- value = output value + burned value + fee + deposits
+-- value + withdrawn value = output value + burned value + fee + deposits
 computeBalancedTxSkel :: (MonadBlockChainBalancing m) => Wallet -> BalancingOutputs -> TxSkel -> Fee -> m TxSkel
 computeBalancedTxSkel balancingWallet balancingUtxos txSkel@TxSkel {..} (Script.lovelace -> feeValue) = do
+  params <- getParams
   -- We compute the necessary values from the skeleton that are part of the
   -- equation, except for the `feeValue` which we already have.
   let (burnedValue, mintedValue) = Api.split $ txSkelMintsValue txSkelMints
       outValue = txSkelValueInOutputs txSkel
+      withdrawnValue = txSkelWithdrawnValue txSkel
   inValue <- txSkelInputValue txSkel
   depositedValue <- toValue <$> txSkelProposalsDeposit txSkel
   -- We compute the values missing in the left and right side of the equation
-  let (missingRight, missingLeft) = Api.split $ outValue <> burnedValue <> feeValue <> depositedValue <> PlutusTx.negate (inValue <> mintedValue)
+  let (missingRight, missingLeft) = Api.split $ outValue <> burnedValue <> feeValue <> depositedValue <> PlutusTx.negate (inValue <> mintedValue <> withdrawnValue)
+  -- We compute the minimal ada requirement of the missing payment
+  rightMinAda <- case getTxSkelOutMinAda params $ paysPK balancingWallet missingRight of
+    Left err -> throwError $ MCEGenerationError err
+    Right a -> return a
+  -- We compute the current ada of the missing payment. If the missing payment
+  -- is not empty and the minimal ada is not present, some value is missing.
+  let Script.Lovelace rightAda = missingRight ^. Script.adaL
+      missingAda = rightMinAda - rightAda
+      missingAdaValue = if missingRight /= mempty && missingAda > 0 then Script.lovelace missingAda else mempty
+  -- The actual missing value on the left might needs to account for any missing
+  -- min ada on the missing payment of the transaction skeleton. This also has
+  -- to be repercuted on the missing value on the right.
+  let missingLeft' = missingLeft <> missingAdaValue
+      missingRight' = missingRight <> missingAdaValue
   -- This gives us what we need to run our `reachValue` algorithm and append to
   -- the resulting values whatever payment was missing in the initial skeleton
-  let candidatesRaw = second (<> missingRight) <$> reachValue balancingUtxos missingLeft (toInteger $ length balancingUtxos)
+  let candidatesRaw = second (<> missingRight') <$> reachValue balancingUtxos missingLeft' (toInteger $ length balancingUtxos)
   -- We prepare a possible balancing error with the difference between the
   -- requested amount and the maximum amount provided by the balancing wallet
   let totalValue = mconcat $ Api.txOutValue . snd <$> balancingUtxos
-      difference = snd $ Api.split $ missingLeft <> PlutusTx.negate totalValue
+      difference = snd $ Api.split $ missingLeft' <> PlutusTx.negate totalValue
       balancingError = MCEUnbalanceable balancingWallet difference txSkel
   -- Which one of our candidates should be picked depends on three factors
   -- - Whether there exists a perfect candidate set with empty surplus value