+++ tests/Cooked/ReferenceInputsSpec.hs
@@ -10,21 +10,24 @@ import Plutus.Script.Utils.V3.Typed.Scripts qualified as Script
 import Plutus.Script.Utils.Value qualified as Script
 import PlutusLedgerApi.V3 qualified as Api
 import PlutusTx qualified
+import PlutusTx.AssocMap qualified as PlutusTx (lookup)
+import PlutusTx.Eq qualified as PlutusTx
 import PlutusTx.Prelude qualified as PlutusTx
 import Prettyprinter qualified as PP
 import Test.Tasty qualified as Tasty
 import Test.Tasty.HUnit qualified as Tasty
 
--- Foo and Bar are two dummy scripts to test reference inputs. They
--- serve no purpose and make no real sense.
+-- Foo, Bar and Baz are dummy scripts to test reference inputs. They serve no
+-- purpose and make no real sense.
 --
--- Foo contains a pkh in its datum. It can only be spent by ANOTHER
--- public key.
+-- Foo contains a pkh in its datum. It can only be spent by ANOTHER public key.
 --
--- Bar has no datum nor redeemer. Its outputs can only be spent by a
--- public key who can provide a Foo UTxO containing its pkh as
--- reference input (that is a UTxO they could not actually spend,
--- according the the design of Foo).
+-- Bar has no datum nor redeemer. Its outputs can only be spent by a public key
+-- who can provide a Foo UTxO containing its pkh as reference input (that is a
+-- UTxO they could not actually spend, according the the design of Foo).
+--
+-- Baz has no datum nor redeemer. Its outputs can only be spent when a reference
+-- input is provided with a hashed datum contain the integer 10.
 --
 -- The datum in Foo outputs in expected to be inlined.
 
@@ -58,12 +61,6 @@ fooTypedValidator =
         $$(PlutusTx.compile [||fooValidator||])
         $$(PlutusTx.compile [||wrap||])
 
-data Bar
-
-instance Script.ValidatorTypes Bar where
-  type RedeemerType Bar = ()
-  type DatumType Bar = ()
-
 -- | Outputs can only be spent by pks who provide a reference input to
 -- a Foo in which they are mentioned (in an inlined datum).
 barValidator :: () -> () -> Api.ScriptContext -> Bool
@@ -77,11 +74,33 @@ barValidator _ _ (Api.ScriptContext txInfo _) =
         Just (FooDatum pkh) -> PlutusTx.elem pkh (Api.txInfoSignatories txInfo)
     f _ = False
 
-barTypedValidator :: Script.TypedValidator Bar
+barTypedValidator :: Script.TypedValidator MockContract
 barTypedValidator =
   let wrap = Script.mkUntypedValidator
-   in Script.mkTypedValidator @Bar
+   in Script.mkTypedValidator @MockContract
         $$(PlutusTx.compile [||barValidator||])
+        $$(PlutusTx.compile [||wrap||])
+
+bazValidator :: () -> () -> Api.ScriptContext -> Bool
+bazValidator _ _ context =
+  let info = Api.scriptContextTxInfo context
+      refInputs = Api.txInfoReferenceInputs info
+      txData = Api.txInfoData info
+   in case refInputs of
+        [myRefInput] ->
+          let Api.TxOut _ _ dat _ = Api.txInInfoResolved myRefInput
+           in case dat of
+                (Api.OutputDatumHash hash) -> case PlutusTx.lookup hash txData of
+                  Nothing -> False
+                  Just (Api.Datum a) -> PlutusTx.unsafeFromBuiltinData @Integer a PlutusTx.== 10
+                _ -> False
+        _ -> False
+
+bazTypedValidator :: Script.TypedValidator MockContract
+bazTypedValidator =
+  let wrap = Script.mkUntypedValidator
+   in Script.mkTypedValidator @MockContract
+        $$(PlutusTx.compile [||bazValidator||])
         $$(PlutusTx.compile [||wrap||])
 
 trace1 :: (MonadBlockChain m) => m ()
@@ -104,8 +123,30 @@ trace1 = do
           txSkelSigners = [wallet 3]
         }
 
+trace2 :: (MonadBlockChain m) => m ()
+trace2 = do
+  refORef : scriptORef : _ <-
+    validateTxSkel'
+      ( txSkelTemplate
+          { txSkelOuts =
+              [ paysPK (wallet 1) (Script.ada 2) `withDatum` (10 :: Integer),
+                paysScript bazTypedValidator () (Script.ada 10)
+              ],
+            txSkelSigners = [wallet 2]
+          }
+      )
+  void $
+    validateTxSkel $
+      txSkelTemplate
+        { txSkelSigners = [wallet 1],
+          txSkelIns = Map.singleton scriptORef (txSkelSomeRedeemer ()),
+          txSkelInsReference = Set.singleton refORef
+        }
+
 tests :: Tasty.TestTree
 tests =
   Tasty.testGroup
     "Reference inputs"
-    [Tasty.testCase "Can reference an input that can't be spent" (testSucceeds def trace1)]
+    [ Tasty.testCase "We can reference an input that can't be spent" (testSucceeds def trace1),
+      Tasty.testCase "We can decode the datum hash from a reference input" (testSucceeds def trace2)
+    ]